pyproject.toml
[tool.poetry]
name = "visibleone"
version = "0.1.0"
description = ""
authors = ["Win Htet Aung"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
graphene = "^3.2.2"
graphene-django = "^3.1.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
